chore: continue services refacto for reflection methods
