+++ docs/src/Services/Reference/Reflection/ReflectionHelper.php
@@ -25,7 +25,8 @@ class ReflectionHelper
     public function __construct(
         private readonly PhpDocHelper $phpDocHelper,
         private readonly OutputFormatter $outputFormatter,
-        private readonly ReflectionPropertyHelper $propertyHelper
+        private readonly ReflectionPropertyHelper $propertyHelper,
+        private readonly ReflectionMethodHelper $methodHelper
     ) {
     }
 
@@ -138,7 +139,7 @@ class ReflectionHelper
     {
         $methods = [];
         foreach ($reflectionClass->getMethods() as $method) {
-            if (!$this->methodHasToBeSkipped($method, $reflectionClass)) {
+            if (!$this->methodHelper->methodHasToBeSkipped($method, $reflectionClass)) {
                 $methods[] = $method;
             }
         }
@@ -149,7 +150,7 @@ class ReflectionHelper
         $content .= '## Methods: '.\PHP_EOL;
 
         foreach ($methods as $method) {
-            $typedParameters = $this->getParametersWithType($method);
+            $typedParameters = $this->methodHelper->getParametersWithType($method);
 
             $content .= "<a className=\"anchor\" href=\"#{$method->getName()}\" id=\"{$method->getName()}\">§</a>".\PHP_EOL;
 
@@ -160,7 +161,7 @@ class ReflectionHelper
                 .'( '
                 .implode(', ', $typedParameters)
                 .' ): '
-                .$this->outputFormatter->addCssClasses($this->getReturnType($method), ['token', 'keyword'])
+                .$this->outputFormatter->addCssClasses($this->methodHelper->getReturnType($method), ['token', 'keyword'])
                 .\PHP_EOL;
 
             $phpDoc = $this->phpDocHelper->getPhpDoc($method);
@@ -188,102 +189,6 @@ class ReflectionHelper
         }
 
         return $content;
-    }
-
-    private function isConstruct(\ReflectionMethod $method): bool
-    {
-        return '__construct' === $method->getName();
-    }
-
-    private function isAccessor(\ReflectionMethod $method): bool
-    {
-        foreach ($method->getDeclaringClass()->getProperties() as $property) {
-            if (str_contains($method->getName(), ucfirst($property->getName()))) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    /**
-     * Checks if a method is actually from a Trait or an extended class.
-     */
-    private function isFromExternalClass(\ReflectionMethod $method, \ReflectionClass $class): bool
-    {
-        return $method->getFileName() !== $class->getFileName();
-    }
-
-    private function methodHasToBeSkipped(\ReflectionMethod $method, \ReflectionClass $reflectionClass): bool
-    {
-        return $this->isFromExternalClass($method, $reflectionClass)
-            || str_contains($this->getModifier($method), 'private')
-            || $this->isAccessor($method)
-            || $this->isConstruct($method);
-    }
-
-    /**
-     * @return array<string, string>
-     */
-    private function getParametersWithType(\ReflectionMethod $method): array
-    {
-        $typedParameters = [];
-        foreach ($method->getParameters() as $parameter) {
-            $parameterName = $this->getParameterName($parameter);
-            $type = $parameter->getType();
-            if (!$type) {
-                $typedParameters[] = $this->outputFormatter->addCssClasses($parameterName, ['token', 'variable']).$this->getDefaultValueString($parameter);
-                continue;
-            }
-            if ($type instanceof \ReflectionUnionType) {
-                $namedTypes = array_map(function (\ReflectionNamedType $namedType) {
-                    return $this->outputFormatter->linkClasses($namedType);
-                }, $type->getTypes());
-
-                $typedParameters[] = implode('|', $namedTypes).' ReflectionHelper.php'.$this->outputFormatter->addCssClasses($parameterName, ['token', 'variable']).$this->getDefaultValueString($parameter);
-            }
-            if ($type instanceof \ReflectionIntersectionType) {
-                $namedTypes = array_map(function (\ReflectionNamedType $namedType) {
-                    return $this->outputFormatter->linkClasses($namedType);
-                }, $type->getTypes());
-
-                $typedParameters[] = implode('&', $namedTypes).' ReflectionHelper.php'.$this->outputFormatter->addCssClasses($parameterName, ['token', 'variable']).$this->getDefaultValueString($parameter);
-            }
-            if ($type instanceof \ReflectionNamedType) {
-                $typedParameters[] = $this->outputFormatter->linkClasses($type).' ReflectionHelper.php'.$this->outputFormatter->addCssClasses($parameterName, ['token', 'variable']).$this->getDefaultValueString($parameter);
-            }
-        }
-
-        return $typedParameters;
-    }
-
-    private function getParameterName(\ReflectionParameter $parameter): string
-    {
-        return $parameter->isPassedByReference() ? '&$'.$parameter->getName() : '$'.$parameter->getName();
-    }
-
-    private function getReturnType(\ReflectionMethod $method): string
-    {
-        $type = $method->getReturnType();
-
-        if (!$type) {
-            return '';
-        }
-
-        if ($type instanceof \ReflectionUnionType) {
-            return implode('|', array_map(function (\ReflectionNamedType $reflectionNamedType): string {
-                return $this->outputFormatter->linkClasses($reflectionNamedType);
-            }, $type->getTypes()
-            ));
-        }
-        if ($type instanceof \ReflectionIntersectionType) {
-            return implode('&', array_map(function (\ReflectionNamedType $reflectionNamedType): string {
-                return $this->outputFormatter->linkClasses($reflectionNamedType);
-            }, $type->getTypes()
-            ));
-        }
-
-        return $this->outputFormatter->linkClasses($type);
     }
 
     public function containsOnlyPrivateMethods(\ReflectionClass $reflectionClass): bool
